IDETECT-725: fixed a concurrency issue; refactoring
